+++ test/tests/test_profile.py
@@ -7,6 +7,7 @@ import importlib
 import logging
 import os
 import shutil
+import pytest
 from subprocess import call
 
 import numpy as np
@@ -1370,6 +1371,7 @@ def test_profile_21(BTO):
     assert len(sdb[sdb['ref_base'] == 'c']) == 0
 
 # This is from test_compare.py
+@pytest.mark.skip(reason="Only run this as-needed")
 def test_UPDATE_COMPARE_TEST_DATA(BTO):
     """
     Run inStrain on bam1 and bam2, and store the results where IS1 and IS2 are
@@ -1377,8 +1379,7 @@ def test_UPDATE_COMPARE_TEST_DATA(BTO):
 
     # Run the first one
     base1 = BTO.test_dir + os.path.basename(BTO.IS1)
-    cmd = "inStrain profile {0} {1} -o {2} --skip_plot_generation".format(BTO.bam1, BTO.fasta,
-                                                                          base1)
+    cmd = f"inStrain profile {BTO.bam1} {BTO.fasta} -o {base1} -g {BTO.genes} --skip_plot_generation"
     print(cmd)
     code = call(cmd, shell=True)
     assert code == 0, code
@@ -1387,17 +1388,16 @@ def test_UPDATE_COMPARE_TEST_DATA(BTO):
     if os.path.isdir(BTO.IS1):
         shutil.rmtree(BTO.IS1)
     shutil.copytree(base1, BTO.IS1)
-    #
-    # # Run the second one
-    # base2 = BTO.test_dir + os.path.basename(BTO.IS2)
-    # cmd = "inStrain profile {0} {1} -o {2} --skip_plot_generation".format(BTO.bam2, BTO.fasta,
-    #                                                                       base2)
-    # print(cmd)
-    # code = call(cmd, shell=True)
-    # assert code == 0, code
-    #
-    # # Copy to new location
-    # if os.path.isdir(BTO.IS2):
-    #     shutil.rmtree(BTO.IS2)
-    # shutil.copytree(base2, BTO.IS2)
-    #
+
+    # Run the second one
+    base2 = BTO.test_dir + os.path.basename(BTO.IS2)
+    cmd = f"inStrain profile {BTO.bam2} {BTO.fasta} -o {base2} -g {BTO.genes} --skip_plot_generation"
+    print(cmd)
+    code = call(cmd, shell=True)
+    assert code == 0, code
+
+    # Copy to new location
+    if os.path.isdir(BTO.IS2):
+        shutil.rmtree(BTO.IS2)
+    shutil.copytree(base2, BTO.IS2)
+